pyproject.toml
[project]
name = "aws-mcp"
version = "0.1.0"
description = "AWS MCP Server for Claude Desktop - provides AWS service access through the Model Context Protocol"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
dependencies = [
    "mcp>=1.3.0",
    "boto3>=1.35.0",
    "botocore>=1.35.0",
    "opentelemetry-api>=1.25.0",
    "opentelemetry-sdk>=1.25.0",
    "opentelemetry-exporter-otlp>=1.25.0",
]

[[project.authors]]
name = "Marcelo"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
aws-mcp = "aws_mcp:run"

[tool.hatch.build.targets.wheel]
packages = ["aws_mcp"]
